✨ feat: Processing the function that receives access tokens as a common utility function
Ref #51
